+++ src/Readarr.Api.V1/Books/MergeBooksResource.cs
@@ -1,0 +1,8 @@
+namespace Readarr.Api.V1.Books
+{
+    public class MergeBooksResource
+    {
+        public int WinnerBookId { get; set; }
+        public int LoserBookId { get; set; }
+    }
+}